git-svn-id: svn+ssh://phpbee.org/home/pahomov/www/phpbee.org/svn/phpbee.org@233 95ad4591-2629-e011-b8a3-0018f302bd40

+++ libs/indexator.lib.php
@@ -70,10 +70,12 @@ class b_tree {
 	private function del_value($path,$value) {
 		$value.=PHP_EOL;
 		$path=$path.DIRECTORY_SEPARATOR.'idx';
-		$a=array_flip(file($path));
-		unset($a[$value]);
-		unset($a['!'.$value]);
-		file_put_contents($path,implode("",array_keys($a)));
+		if (file_exists($path)) {
+			$a=array_flip(file($path));
+			unset($a[$value]);
+			unset($a['!'.$value]);
+			file_put_contents($path,implode("",array_keys($a)));
+		}
 	}
 	
 	private function int2str($int) {
